# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Find your ideal room and your ideal mate. So, let´s roomie",
    "version": "1.0.0",
    "title": "API-LetsRoomie"
  },
  "host": "api-letsroomie.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/profile/{profileId}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get a specific profie",
        "operationId": "getOneProfile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "type": "string",
            "name": "access-token",
            "required": true
          },
          {
            "in": "path",
            "type": "string",
            "name": "profileId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/getProfile"
            }
          },
          "401": {
            "description": "Invalid or missing token"
          }
        }
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Deletes a profile",
        "operationId": "deleteProfileById",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "description": "Id of profile to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/userDeleted"
            }
          },
          "500": {
            "description": "Error updating user"
          }
        }
      }
    }
  },
  "definitions": {
...
    "userDeleted": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "body": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string"
            },
            "data": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "uid": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "__v": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
    "getProfile": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "body": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "places": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "avatar": {
              "type": "string"
            },
            "isHost": {
              "type": "boolean"
            },
            "about": {
              "type": "string"
            },
            "user_id": {
              "type": "string"
            },
            "favorites": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "_v": {
              "type": "integer"
            }
          }
        }
      }
    },
...
  },
...
}